# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "新能源监控系统RESTful API接口文档，提供区域管理、厂站管理、设备管理、采集点管理、告警管理、数据查询等功能。",
        "title": "新能源监控系统 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API支持团队",
            "url": "http://www.new-energy-monitoring.com/support",
            "email": "support@new-energy-monitoring.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/alarms": {
            "get": {
                "description": "获取所有告警的列表，支持分页和过滤",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "告警管理"
                ],
                "summary": "获取告警列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "厂站ID",
                        "name": "station_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "设备ID",
                        "name": "device_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "告警级别",
                        "name": "level",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "状态",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "告警类型",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "开始时间戳",
                        "name": "start_time",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "结束时间戳",
                        "name": "end_time",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.PagedResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.AlarmResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/alarms/{id}": {
            "get": {
                "description": "根据ID获取告警详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "告警管理"
                ],
                "summary": "获取告警详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "告警ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.AlarmResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/alarms/{id}/ack": {
            "put": {
                "description": "确认指定告警",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "告警管理"
                ],
                "summary": "确认告警",
                "parameters": [
                    {
                        "type": "string",
                        "description": "告警ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "确认信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AckAlarmRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.AlarmResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/alarms/{id}/clear": {
            "put": {
                "description": "清除指定告警",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "告警管理"
                ],
                "summary": "清除告警",
                "parameters": [
                    {
                        "type": "string",
                        "description": "告警ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "清除信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ClearAlarmRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.AlarmResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.AckAlarmRequest": {
            "type": "object",
            "required": [
                "operator"
            ],
            "properties": {
                "comment": {
                    "type": "string",
                    "example": "已确认，正在处理"
                },
                "operator": {
                    "type": "string",
                    "example": "admin"
                }
            }
        },
        "dto.AlarmResponse": {
            "type": "object",
            "properties": {
                "acknowledged_at": {
                    "type": "string"
                },
                "acknowledged_by": {
                    "type": "string",
                    "example": "admin"
                },
                "cleared_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "device_id": {
                    "type": "string",
                    "example": "device-001"
                },
                "id": {
                    "type": "string",
                    "example": "alarm-001"
                },
                "level": {
                    "type": "integer",
                    "example": 3
                },
                "message": {
                    "type": "string",
                    "example": "逆变器温度为87.5°C，已超过85°C阈值"
                },
                "point_id": {
                    "type": "string",
                    "example": "point-001"
                },
                "station_id": {
                    "type": "string",
                    "example": "station-001"
                },
                "status": {
                    "type": "integer",
                    "example": 1
                },
                "threshold": {
                    "type": "number",
                    "example": 85
                },
                "title": {
                    "type": "string",
                    "example": "逆变器温度过高"
                },
                "triggered_at": {
                    "type": "string"
                },
                "type": {
                    "type": "string",
                    "example": "limit"
                },
                "value": {
                    "type": "number",
                    "example": 87.5
                }
            }
        },
...
        "dto.ClearAlarmRequest": {
            "type": "object",
            "required": [
                "operator"
            ],
            "properties": {
                "comment": {
                    "type": "string",
                    "example": "问题已解决"
                },
                "operator": {
                    "type": "string",
                    "example": "admin"
                }
            }
        },
...
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "请求参数错误"
                },
                "timestamp": {
                    "type": "integer",
                    "example": 1709500800000
                }
            }
        },
...
        "dto.PagedResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "success"
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "page_size": {
                    "type": "integer",
                    "example": 20
                },
                "timestamp": {
                    "type": "integer",
                    "example": 1709500800000
                },
                "total": {
                    "type": "integer",
                    "example": 100
                }
            }
        },
...
        "dto.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "success"
                },
                "timestamp": {
                    "type": "integer",
                    "example": 1709500800000
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT认证令牌，格式: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}